requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2.5
django-ckeditor==6.7.2
django-import-export==3.3.1
django-js-asset==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-simpleui==2023.8.28
et_xmlfile==2.0.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.5
pillow==11.0.0
pip==24.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.3
tablib==3.5.0
typing_extensions==4.12.2
wheel==0.44.0
xlrd==2.0.1
xlwt==1.3.0
minio==7.2.14
